+++ mapstory/tests/testInitiatives.py
@@ -1,0 +1,10 @@
+from .MapStoryTestMixin import MapStoryTestMixin
+
+
+class InitiativeTests(MapStoryTestMixin):
+    def setUp(self):
+        print("Hello")
+
+    def test_access_permissions(self):
+        print("Hello")
+
